+++ src/pydidas/gui/frames/builders/image_math_frame_builder.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2024 - 2025, Helmholtz-Zentrum Hereon
+# Copyright 2024 - 2026, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ the ImageMathFrame with widgets.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2024 - 2025, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2024 - 2026, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -33,7 +33,7 @@ from typing import Any
 from qtpy import QtCore, QtGui, QtWidgets
 
 from pydidas.core import constants
-from pydidas.core.constants import FONT_METRIC_WIDE_CONFIG_WIDTH
+from pydidas.core.constants import FONT_METRIC_CONSOLE_WIDTH
 from pydidas.widgets import ScrollArea
 from pydidas.widgets.selection import SelectDataFrameWidget
 from pydidas.widgets.silx_plot import PydidasPlot2D
@@ -52,25 +52,25 @@ FLOAT_VALIDATOR.setLocale(LOCAL_SETTINGS)
 _kwargs_for_single_char_label = {
     "bold": True,
     "fontsize_offset": 1,
-    "font_metric_width_factor": 0.7,
+    "font_metric_width_factor": 2,
     "gridPos": (0, -1, 1, 1),
     "margin": 2,
     "minimum_width": 12,
 }
 
 IMAGE_BUFFER_SIZE = 3
-_INPUT_IMAGES = tuple(f"Input image #{i}" for i in range(1, IMAGE_BUFFER_SIZE + 1))
-_IMAGES = tuple(f"Image #{i}" for i in range(1, IMAGE_BUFFER_SIZE + 1))
-_ALL_IMAGES = ("Opened file",) + _INPUT_IMAGES + _IMAGES
-_CURRENT_IMAGES = ("Current image",) + _INPUT_IMAGES + _IMAGES
+_INPUT_IMAGES = [f"Input image #{i}" for i in range(1, IMAGE_BUFFER_SIZE + 1)]
+_IMAGES = [f"Image #{i}" for i in range(1, IMAGE_BUFFER_SIZE + 1)]
+_ALL_IMAGES = ["Opened file"] + _INPUT_IMAGES + _IMAGES
+_CURRENT_IMAGES = ["Current image"] + _INPUT_IMAGES + _IMAGES
 
-IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
+IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any, ...] | dict[str, Any]]] = [
     [
         "create_label",
         ("title", "Image mathematics"),
         {
             "bold": True,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH,
             "fontsize_offset": 4,
             "gridPos": (0, 0, 1, 1),
         },
@@ -79,7 +79,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_empty_widget",
         ("left_container",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH,
             "minimumHeight": 600,
             "parent_widget": None,
         },
@@ -106,7 +106,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("file_selector", SelectDataFrameWidget),
         {
             "import_reference": "ImageMathFrame__image_import",
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH,
             "parent_widget": "left_container",
         },
     ],
@@ -120,7 +120,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_label",
         ("label_store_input_image", "Store loaded image as "),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 2,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH / 2,
             "parent_widget": "store_input_image",
             "sizePolicy": constants.POLICY_MIN_MIN,
             "wordWrap": False,
@@ -130,7 +130,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("combo_store_input_image",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH / 3,
             "gridPos": (0, 1, 1, 1),
             "items": _INPUT_IMAGES,
             "parent_widget": "store_input_image",
@@ -164,7 +164,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("label_display_from_buffer", "Display image:"),
         {
             "bold": True,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH / 3,
             "parent_widget": "display_from_buffer",
             "sizePolicy": constants.POLICY_MIN_MIN,
             "wordWrap": False,
@@ -185,7 +185,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("combo_display_image",),
         {
             "alignment": constants.ALIGN_TOP_RIGHT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 2,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH / 2,
             "font_metric_height_factor": 1,
             "gridPos": (0, 2, 1, 1),
             "items": _ALL_IMAGES,
@@ -214,7 +214,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("ops_arithmetic_target",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 4.8,
+            "font_metric_width_factor": 17,
             "gridPos": (0, 0, 1, 1),
             "items": _IMAGES,
             "parent_widget": "ops_arithmetic",
@@ -230,7 +230,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("combo_ops_arithmetic_input",),
         {
             "alignment": constants.ALIGN_CENTER_LEFT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3.2,
+            "font_metric_width_factor": 23,
             "gridPos": (0, -1, 1, 1),
             "items": _CURRENT_IMAGES,
             "parent_widget": "ops_arithmetic",
@@ -240,7 +240,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("combo_ops_arithmetic_operation",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 10,
+            "font_metric_width_factor": 10,
             "gridPos": (0, -1, 1, 1),
             "items": ["+", "-", "/", "x"],
             "minimum_width": 35,
@@ -251,7 +251,6 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_lineedit",
         ("io_ops_arithmetic_input",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 4.8,
             "font_metric_height_factor": 1,
             "gridPos": (0, -1, 1, 1),
             "parent_widget": "ops_arithmetic",
@@ -259,21 +258,11 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         },
     ],
     [
-        "create_spacer",
-        ("spacer_arithmetic_end",),
-        {
-            "gridPos": (0, -1, 1, 1),
-            "parent_widget": "ops_arithmetic",
-            "policy": QtWidgets.QSizePolicy.Expanding,
-            "vertical_policy": QtWidgets.QSizePolicy.Fixed,
-        },
-    ],
-    [
         "create_button",
         ("but_ops_arithmetic_execute", "Apply arithmetic operation"),
         {
             "alignment": constants.ALIGN_TOP_RIGHT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 2.25,
+            "font_metric_width_factor": 36,
             "parent_widget": "left_container",
         },
     ],
@@ -292,7 +281,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("ops_operator_target",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 4.8,
+            "font_metric_width_factor": 17,
             "gridPos": (0, 0, 1, 1),
             "items": _IMAGES,
             "parent_widget": "ops_operator",
@@ -308,7 +297,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("ops_operator_func",),
         {
             "currentIndex": 4,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 4.8,
+            "font_metric_width_factor": 17,
             "gridPos": (0, -1, 1, 1),
             "items": UFUNCS,
             "parent_widget": "ops_operator",
@@ -324,7 +313,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("combo_ops_operator_input",),
         {
             "alignment": constants.ALIGN_CENTER_LEFT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3.5,
+            "font_metric_width_factor": 23,
             "gridPos": (0, -1, 1, 1),
             "items": _CURRENT_IMAGES,
             "parent_widget": "ops_operator",
@@ -343,7 +332,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_lineedit",
         ("io_ops_operator_input",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 9,
+            "font_metric_width_factor": FONT_METRIC_CONSOLE_WIDTH / 9,
             "gridPos": (0, -1, 1, 1),
             "parent_widget": "ops_operator",
             "validator": FLOAT_VALIDATOR,
@@ -370,7 +359,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("but_ops_operator_execute", "Apply operator"),
         {
             "alignment": constants.ALIGN_TOP_RIGHT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 2.25,
+            "font_metric_width_factor": 36,
             "parent_widget": "left_container",
         },
     ],
@@ -389,7 +378,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("ops_image_arithmetic_target",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 4.8,
+            "font_metric_width_factor": 17,
             "items": _IMAGES,
             "parent_widget": "ops_image_arithmetic",
         },
@@ -404,7 +393,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("combo_ops_image_arithmetic_input_1",),
         {
             "alignment": constants.ALIGN_CENTER_LEFT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3.5,
+            "font_metric_width_factor": 23,
             "gridPos": (0, -1, 1, 1),
             "items": _CURRENT_IMAGES,
             "parent_widget": "ops_image_arithmetic",
@@ -414,7 +403,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         "create_combo_box",
         ("combo_ops_image_arithmetic_operation",),
         {
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 10,
+            "font_metric_width_factor": 10,
             "minimum_width": 35,
             "gridPos": (0, -1, 1, 1),
             "items": ["+", "-", "/", "x"],
@@ -426,20 +415,9 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("combo_ops_image_arithmetic_input_2",),
         {
             "alignment": constants.ALIGN_CENTER_LEFT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 3.5,
             "gridPos": (0, -1, 1, 1),
             "items": _CURRENT_IMAGES,
             "parent_widget": "ops_image_arithmetic",
-        },
-    ],
-    [
-        "create_spacer",
-        ("spacer_arithmetic_end",),
-        {
-            "gridPos": (0, -1, 1, 1),
-            "parent_widget": "ops_arithmetic",
-            "policy": QtWidgets.QSizePolicy.Expanding,
-            "vertical_policy": QtWidgets.QSizePolicy.Fixed,
         },
     ],
     [
@@ -447,7 +425,7 @@ IMAGE_MATH_FRAME_BUILD_CONFIG: list[list[str | tuple[Any] | dict[str, Any]]] = [
         ("but_ops_image_arithmetic_execute", "Apply arithmetic operation"),
         {
             "alignment": constants.ALIGN_TOP_RIGHT,
-            "font_metric_width_factor": FONT_METRIC_WIDE_CONFIG_WIDTH / 2.25,
+            "font_metric_width_factor": 36,
             "parent_widget": "left_container",
         },
     ],